added progress on create command using the workflow report
