+++ pkg/cli/admin/nodeimage/monitor.go
@@ -161,7 +161,7 @@ func (o *MonitorOptions) Run() error {
 
 	podName := o.nodeJoinerPod.GetName()
 
-	if err := o.waitForContainerRunning(ctx); err != nil {
+	if err := o.waitForRunningPod(ctx); err != nil {
 		klog.Errorf("monitoring did not start: %s", err)
 		return fmt.Errorf("monitoring did not start for pod %s: %s", podName, err)
 	}
